+++ Source/Graphics/Vulkan/VkVertexView.cpp
@@ -1,7 +1,7 @@
 #include "VkVertexView.h"
 
 Eugene::VkVertexView::VkVertexView(size_t size, size_t vertexNum, vk::Buffer& buffer) :
-	VertexView{vertexNum}
+	VertexView{static_cast<std::uint32_t>(vertexNum)}
 {
 	buffer_ = buffer;
 }
